1. All templates now use a consistent "Continue button" partial. This partial now distinguishes between logged-in users and users who have merely completed a request to apply for an account, reset a password, etc. more effectively.
If the user is logged in and sfApplyPlugin_afterLogin is set to a route or URL, The continue button points to that route or URL.

If the user is not logged in, or sfApplyPlugin_afterLogin is not set, the continue button points to the route or URL indicated by sfApplyPlugin_after.

If neither key is set, the button points to @homepage.

2. If a user who has never verified their account attempts to reset their password, they are given another opportunity to verify their account instead of a 404 error.

3. security.yml has been upgraded for Symfony 1.3.



git-svn-id: http://svn.symfony-project.com/plugins/sfDoctrineApplyPlugin/trunk@24892 ee427ae8-e902-0410-961c-c3ed070cd9f9